move around some sink functions
